requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
fastmcp>=2.11.0
pydantic>=2.0.0
python-multipart>=0.0.6
google-genai>=0.8.0
requests>=2.31.0
